==> R/00_doit.R <==
# Run all scripts ---------------------------------------------------------
source(file = "R/01_load.R")
source(file = "R/02_clean.R")
source(file = "R/03_augment.R")
source(file = "R/04_EDA.R")
source(file = "R/05_logistic_reg.R")
source(file = "R/06_clustering.R")
source(file = "R/07_pca.R")


# Knit the Presentation slides --------------------------------------------
rmarkdown::render('doc/Presentation.Rmd',output_file='Presentation.html')
